KS-1-4 Update variable types

+++ src/main/kotlin/lesson1/task4.kt
@@ -1,9 +1,9 @@
 fun main() {
     val length: Long  = 40_868_600_000 // мм расстояние
-    val age: Int = 27 // возраст
-    val partOfDay: Double = 0.075 // часть дня
-    val seconds: Short = 6480 // секунд провел
-    val partOfYear: Double = 2.0547945205479453E-4 // часть года
+    val age: Byte = 27 // возраст
+    val partOfDay: Float = 0.075F // часть дня
+    val seconds: Int = 6480 // секунд провел
+    val partOfYear: Float = 2.0547945205479453E-4F // часть года
     val apogee: Int = 327_000 // м апогей орбиты
 
     println("Расстояние: $length мм")
